pyproject.toml
[project]
name = "swli-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.2.1",
    "cupy-cuda12x>=13.3.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "plotly>=6.0.0",
    "pytorch-wavelets>=1.3.0",
    "pywavelets>=1.8.0",
    "rerun-sdk>=0.21.0",
    "scipy>=1.15.1",
    "torch>=2.5.1",
    "transformers>=4.47.1",
]



[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

